Initial GCP Signer implementation
Very bare bones Signer for Google Cloud KMS: Private keys live in KMS,
signing happens in KMS (although payloading hash happens in Signer).

This is not super usable without issue #447 but demonstrates the simplicity.
With 447, the usage pattern nwould be

signer = Signer(
    "gcpkms://projects/openssf/locations/global/keyRings/securesystemslib-test-keyring/cryptoKeys/securesystemslib-test-key/cryptoKeyVersions/1",
    pubkey
)

so all GCP specific details are in the private key URI.

Key creation is not supported at this point.