Initial GCP Signer implementation
Very bare bones Signer for Google Cloud KMS: Private keys live in KMS,
signing happens in KMS (although payload hashing happens in Signer).

This is not super usable without issue 447 but demonstrates the simplicity.

Key creation is not supported at this point.

A test is added with a few caveats:
* dependencies are not added to requirements.txt: this would
  more than triple the size of requirements-pinned.txt...
  Not sure what the best path here is
* Test only works on GitHub (because of the authentication),
  and only on branches within the upstream repo: not on PRs from forks
* Test is run only once: it's a smoke test, not an exhaustive matrix
  test.